Toccata Cleanup - Part 4
* Drop the toccata activation argument from the mining manager and mempool
* Remove the pre-Toccata standard transaction mass cap and its relaxation window
* Make the relay fee floor unconditional and delete the legacy minimum relay fee
* Drop the unused priority and DAA score arguments from the standardness checks
* Reduce the transient mass activation tests to the durable template limit checks
* Remove the mempool mass rejection error variants left without constructors
