+++ mining/src/mempool/validate_and_insert_transaction.rs
@@ -30,7 +30,7 @@ impl Mempool {
         transaction.calculated_non_contextual_masses = Some(consensus.calculate_transaction_non_contextual_masses(&transaction.tx)?);
         let virtual_daa_score = consensus.get_virtual_daa_score();
         self.validate_transaction_limits_in_isolation(&transaction, virtual_daa_score)?;
-        self.validate_transaction_std_in_isolation(&transaction, virtual_daa_score)?;
+        self.validate_transaction_std_in_isolation(&transaction)?;
         let feerate_threshold = self.get_replace_by_fee_constraint(&transaction, rbf_policy, virtual_daa_score)?;
         self.populate_mempool_entries(&mut transaction)?;
         Ok(TransactionPreValidation { transaction, feerate_threshold })
@@ -74,7 +74,7 @@ impl Mempool {
 
         // Perform mempool in-context validations prior to possible RBF replacements
         self.validate_transaction_limits_in_context(&transaction, virtual_daa_score)?;
-        self.validate_transaction_std_in_context(&transaction, priority, virtual_daa_score)?;
+        self.validate_transaction_std_in_context(&transaction)?;
 
         // Check double spends and try to remove them if the RBF policy requires it
         let removed_transaction = self.execute_replace_by_fee(&transaction, rbf_policy, virtual_daa_score)?;
@@ -146,21 +146,16 @@ impl Mempool {
         Ok(())
     }
 
-    fn validate_transaction_std_in_isolation(&self, transaction: &MutableTransaction, virtual_daa_score: u64) -> RuleResult<()> {
+    fn validate_transaction_std_in_isolation(&self, transaction: &MutableTransaction) -> RuleResult<()> {
         if !self.config.accept_non_standard {
-            self.check_transaction_standard_in_isolation(transaction, virtual_daa_score)?;
+            self.check_transaction_standard_in_isolation(transaction)?;
         }
         Ok(())
     }
 
-    fn validate_transaction_std_in_context(
-        &self,
-        transaction: &MutableTransaction,
-        priority: Priority,
-        virtual_daa_score: u64,
-    ) -> RuleResult<()> {
+    fn validate_transaction_std_in_context(&self, transaction: &MutableTransaction) -> RuleResult<()> {
         if !self.config.accept_non_standard {
-            self.check_transaction_standard_in_context(transaction, priority, virtual_daa_score)?;
+            self.check_transaction_standard_in_context(transaction)?;
         }
         Ok(())
     }
